Move endpoints and extractors to separate modules. Simplify url building (replace strings to basic urls).
